prefs initial draft
